Attempt to make it Delphi 10.1 compatible
Of course this can be a long term solution as it is more a kind of hack
as a real solution but in the meantime...

As is seems to work well for fields, I was not able to get a proper
adress for the procedure / functions. This is why style is not applied
everywhere. Perhaps does someone have a better idea.